+++ core/src/main/java/arez/annotations/Computed.java
@@ -89,4 +89,12 @@ public @interface Computed
    * @return true if invocations of the computable must be executed in an {@link arez.Environment}, false otherwise.
    */
   boolean requireEnvironment() default false;
+
+  /**
+   * Return true if the return value of the comptable value should be reported to the Arez spy subsystem.
+   * It is useful to disable reporting for large, circular or just uninteresting parameters to the spy infrastructure.
+   *
+   * @return true to report the return value, false otherwise.
+   */
+  boolean reportResult() default true;
 }