+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 * **\[core\]** Rename the annotation `@ComputedValueRef` to `@ComputableValueRef`.
 * **\[core\]** Fix invalid invariant failure triggered when invoking an action when compile time setting
   `arez.check_invariants` is set to `false` but `arez.check_api_invariants` is set to `true`.
+* **\[core\]** Support the flag `Flags.NO_REPORT_RESULT` being passed to the `ArezContext.computable(...)` method
+  to stop the result of computable calculation being reported in the `ComputeCompletedEvent` spy event.
 
 ### [v0.111](https://github.com/arez/arez/tree/v0.111) (2018-10-19)
 [Full Changelog](https://github.com/arez/arez/compare/v0.110...v0.111)